refactor: externalize CoinjoinPrecondition hook
